Error Fixes #13 - app opens, allows user to sign/register in, all the pages are visible and wont crash the app.

+++ app/src/main/java/com/st10083866/prog7313_poe_personalbudgetapp/ui/loginReg/LoginPageFragment.kt
@@ -57,7 +57,7 @@ class LoginPageFragment : Fragment() {
                         Toast.makeText(requireContext(), "Login successful!", Toast.LENGTH_SHORT).show()
 
                         val session = SessionManager(requireContext())
-                        session.saveUserId(user.userId.toLong())
+                        session.saveUserId(user.userId.toInt())
 
                         val intent = Intent(requireContext(), MainPageActivity::class.java)
                         intent.putExtra("USER_ID", user.userId)
